requirements.txt
# requirements.txt

# Core RAG and Crawling Packages
requests==2.31.0
beautifulsoup4==4.12.3
sentence-transformers==2.6.1
numpy==1.26.4
faiss-cpu==1.8.0          # Required for FAISS vector search in rag.py

# API Framework and Server
fastapi==0.111.0
uvicorn[standard]==0.29.0 # ASGI server to run FastAPI with performance extras

# Environment Variable Management
python-dotenv==1.0.1

#Encryption
cryptography==42.0.5